Fixed base tag in calcapp/index.html
This is necessary because of how urls are resolved. Having a /at the beginning tells the resolver that files are relative to the server root, which for the github.io site is the root where index.html is. When basehref is concatenated with the other src urls, they read "/calcapp/file_name" which is exactly what they are relative to the server location.

See: https://stackoverflow.com/questions/2005079/absolute-vs-relative-urls for a great explanation.